Fallback to default locale if ru_RU.KOI8-R is not found.
Fix Travis and gitpod initialization. Whitespace-only change to main.cpp.

+++ src/lang/stringlist.cpp
@@ -34,7 +34,7 @@ bool StringList::superListOf(const StringList& smallerList) const
             return false;
     }
 
-    return smallerList.size() == matchesFound;
+    return (int)smallerList.size() == matchesFound;
 }
 
 DLString StringList::join(const DLString &delim) const
